resolve linter for morpheusvm
