+++ cli/spam.go
@@ -18,7 +18,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"golang.org/x/sync/errgroup"
 
@@ -50,39 +49,6 @@ var (
 	inflight atomic.Int64
 	sent     atomic.Int64
 )
-
-type SpamHelper interface {
-	// CreateAccount generates a new account and returns the [PrivateKey].
-	//
-	// The spammer tracks all created accounts and orchestrates the return of funds
-	// sent to any created accounts on shutdown. If the spammer exits ungracefully,
-	// any funds sent to created accounts will be lost unless they are persisted by
-	// the [SpamHelper] implementation.
-	CreateAccount() (*PrivateKey, error)
-	// GetFactory returns the [chain.AuthFactory] for a given private key.
-	//
-	// A [chain.AuthFactory] signs transactions and provides a unit estimate
-	// for using a given private key (needed to estimate fees for a transaction).
-	GetFactory(pk *PrivateKey) (chain.AuthFactory, error)
-
-	// CreateClient instructs the [SpamHelper] to create and persist a VM-specific
-	// JSONRPC client.
-	//
-	// This client is used to retrieve the [chain.Parser] and the balance
-	// of arbitrary addresses.
-	//
-	// TODO: consider making these functions part of the required JSONRPC
-	// interface for the HyperSDK.
-	CreateClient(uri string, networkID uint32, chainID ids.ID) error
-	GetParser(ctx context.Context, chainID ids.ID) (chain.Parser, error)
-	LookupBalance(choice int, address string) (uint64, error)
-
-	// GetTransfer returns a list of actions that sends [amount] to a given [address].
-	//
-	// Memo is used to ensure that each transaction is unique (even if between the same
-	// sender and receiver for the same amount).
-	GetTransfer(address codec.Address, amount uint64, memo []byte) []chain.Action
-}
 
 func (h *Handler) Spam(sh SpamHelper) error {
 	ctx := context.Background()
@@ -132,7 +98,7 @@ func (h *Handler) Spam(sh SpamHelper) error {
 	}
 
 	// Compute max units
-	parser, err := sh.GetParser(ctx, chainID)
+	parser, err := sh.GetParser(ctx)
 	if err != nil {
 		return err
 	}
